Moved the kmean test
